#67 ... refactored the unit-tests, incorporated retries, and separated the b2cContactProcess test-classes into separate, logically organized classes.
