+++ test/_use-cases/sfdc/b2cCustomerProfileUpdate.live.test.js
@@ -24,11 +24,17 @@ describe('Updating an SFDC Contact representing a B2C Commerce Customer Profile 
     // noinspection JSAccessibilityCheck
     this.timeout(config.get('unitTests.testData.describeTimeout'));
 
+    // Configure the total number of retries supported per test
+    // noinspection JSAccessibilityCheck
+    this.retries(config.get('unitTests.testData.testRetryCount'));
+
     // Initialize local variables
     let environmentDef,
+        disablePurge,
         baseRequest,
         initResults,
         testProfile,
+        testEmail,
         profileUpdate,
         customerListId,
         siteId,
@@ -46,6 +52,9 @@ describe('Updating an SFDC Contact representing a B2C Commerce Customer Profile 
         // Retrieve the runtime environment
         environmentDef = getRuntimeEnvironment();
 
+        // Default the disable purge property
+        disablePurge = config.get('unitTests.testData.disablePurge');
+
         // Default the sleepTimeout to enforce in unit-tests
         sleepTimeout = config.get('unitTests.testData.sleepTimeout');
         purgeSleepTimeout = sleepTimeout / 2;
@@ -54,9 +63,15 @@ describe('Updating an SFDC Contact representing a B2C Commerce Customer Profile 
         customerListId = config.get('unitTests.testData.b2cCustomerList').toString();
         siteId = config.util.toObject(config.get('unitTests.testData.b2cSiteCustomerLists'))[customerListId];
 
+        // Generate a random email to leverage for profiles
+        testEmail = common.getEmailForTestProfile();
+
         // Retrieve the b2c customer profile template that we'll use to exercise this test
         testProfile = config.util.toObject(config.get('unitTests.testData.profileTemplate'));
         profileUpdate = config.util.toObject(config.get('unitTests.testData.updateTemplate'));
+
+        // Update the email address with a random email
+        testProfile.customer.email = testEmail;
 
         // Default the sync-configuration to leverage; sync-on-login and sync-once are enabled
         syncGlobalEnable = config.get('unitTests.b2cCRMSyncConfigManager.base');
@@ -74,6 +89,19 @@ describe('Updating an SFDC Contact representing a B2C Commerce Customer Profile 
 
             // Audit the authorization token for future rest requests
             sfdcAuthCredentials = initResults.multiCloudInitResults.sfdcAuthCredentials;
+
+            // Is the purge disabled?
+            if (disablePurge === true) {
+
+                // Audit to the console that the purge is disabled
+                console.log(' -- disablePurge is enabled: test-data is not cleaned-up after each test or test-run');
+
+            } else {
+
+                // Purge the customer data in B2C Commerce and SFDC
+                await useCaseProcesses.b2cCustomerPurge(b2cAdminAuthToken, sfdcAuthCredentials.conn);
+
+            }
 
         } catch (e) {
 
@@ -104,6 +132,8 @@ describe('Updating an SFDC Contact representing a B2C Commerce Customer Profile 
 
         // Implement a pause to allow the PlatformEvent to fire
         await useCaseProcesses.sleep(sleepTimeout);
+
+        console.log(output.b2cRegisterResults.data);
 
         // Audit the customerNo of the newly registered customer to that we can
         // delete this customer record as part of the tear-down process
